refactor: address code review findings for specificity routing
- Extract keyword lists to keywords.ts (config.ts was 1077 lines)
- Remove unused combinedText import from scan-messages.ts
- Remove unused SpecificityParamDto class
- Remove duplicate 'social ad' keyword
- Remove unused SPECIFICITY_LABELS/DESCRIPTIONS from shared package
