+++ packages/backend/src/scoring/config.ts
@@ -1,955 +1,7 @@
 import { ScorerConfig } from './types';
+import { DEFAULT_KEYWORDS } from './keywords';
 
-export const DEFAULT_KEYWORDS: Record<string, string[]> = {
-  formalLogic: [
-    'prove',
-    'proof',
-    'derive',
-    'derivation',
-    'theorem',
-    'lemma',
-    'corollary',
-    'if and only if',
-    'iff',
-    'contradiction',
-    'contrapositive',
-    'induction',
-    'deduction',
-    'axiom',
-    'postulate',
-    'qed',
-    'formally verify',
-    'satisfiability',
-    'soundness',
-    'completeness',
-    'undecidable',
-    'reducible',
-    'irrational',
-    'rational',
-    'conjecture',
-    'hypothesis',
-  ],
-
-  analyticalReasoning: [
-    'compare',
-    'contrast',
-    'evaluate',
-    'assess',
-    'trade-offs',
-    'tradeoffs',
-    'pros and cons',
-    'advantages and disadvantages',
-    'weigh',
-    'critically analyze',
-    'strengths and weaknesses',
-    'implications',
-    'ramifications',
-    'nuance',
-    'on the other hand',
-    'counterargument',
-    'analyze',
-    'analysis',
-    'explain',
-  ],
-
-  codeGeneration: [
-    'write a function',
-    'implement',
-    'create a class',
-    'build a component',
-    'write code',
-    'write a script',
-    'code this',
-    'program this',
-    'create an api',
-    'build a module',
-    'scaffold',
-    'boilerplate',
-    'write a test',
-    'write tests',
-    'generate code',
-    'component',
-    'endpoint',
-    'handler',
-    'controller',
-    'write a component',
-    'create a component',
-    'build a service',
-    'implementing',
-    'implemented',
-    'implementation',
-    'function',
-    'method',
-    'class',
-    'variable',
-    'array',
-    'object',
-    'loop',
-    'syntax',
-    'compile',
-    'runtime',
-    'parse',
-    'import',
-    'export',
-    'module',
-    'async',
-    'await',
-    'callback',
-    'promise',
-    'return',
-    'parameter',
-    'argument',
-    'code',
-    'coding',
-    'script',
-    'snippet',
-    'console',
-    'print',
-    'output',
-    'input',
-    'string',
-    'integer',
-    'boolean',
-    'null',
-    'undefined',
-    'type',
-    'interface',
-    'enum',
-    'const',
-    'let',
-    'var',
-    'def',
-    'lambda',
-    'for loop',
-    'while loop',
-    'if else',
-    'try catch',
-    'error handling',
-    'unit test',
-  ],
-
-  codeReview: [
-    'fix this bug',
-    'debug',
-    'why does this fail',
-    'review this code',
-    "what's wrong with",
-    'code review',
-    'refactor',
-    'optimize this code',
-    'find the error',
-    'stack trace',
-    'exception',
-    'segfault',
-    'memory leak',
-    'race condition',
-    'deadlock',
-    'off by one',
-    'typeerror',
-    'referenceerror',
-    'syntaxerror',
-    'vulnerabilities',
-    'vulnerability',
-  ],
-
-  technicalTerms: [
-    'algorithm',
-    'kubernetes',
-    'distributed',
-    'microservice',
-    'database',
-    'architecture',
-    'infrastructure',
-    'deployment',
-    'pipeline',
-    'middleware',
-    'encryption',
-    'authentication',
-    'authorization',
-    'latency',
-    'throughput',
-    'concurrency',
-    'parallelism',
-    'serialization',
-    'deserialization',
-    'react',
-    'graphql',
-    'typescript',
-    'docker',
-    'sql',
-    'redis',
-    'frontend',
-    'backend',
-    'server',
-    'api',
-    'oauth',
-    'repositories',
-    'repository',
-    'security',
-    'python',
-    'javascript',
-    'java',
-    'golang',
-    'rust',
-    'ruby',
-    'swift',
-    'kotlin',
-    'php',
-    'html',
-    'css',
-    'node.js',
-    'nodejs',
-    'nextjs',
-    'vue',
-    'angular',
-    'svelte',
-    'solidjs',
-    'django',
-    'flask',
-    'fastapi',
-    'express',
-    'nestjs',
-    'postgres',
-    'mongodb',
-    'webpack',
-    'vite',
-    'npm',
-    'git',
-    'cli',
-    'json',
-    'yaml',
-    'regex',
-  ],
-
-  simpleIndicators: [
-    'what is',
-    'define',
-    'translate',
-    'thanks',
-    'thank you',
-    'yes',
-    'no',
-    'ok',
-    'okay',
-    'sure',
-    'got it',
-    'hi',
-    'hello',
-    'hey',
-    'bye',
-    'goodbye',
-    'how are you',
-    'good morning',
-    'good night',
-    'please',
-    'help',
-    'what time',
-    'where is',
-    'who is',
-  ],
-
-  multiStep: [
-    'first',
-    'then',
-    'after that',
-    'finally',
-    'step 1',
-    'step 2',
-    'step 3',
-    'next',
-    'subsequently',
-    'once you',
-    'followed by',
-    'in sequence',
-    'phase 1',
-    'phase 2',
-    'stage 1',
-    'stage 2',
-    'workflow',
-    'pipeline',
-  ],
-
-  creative: [
-    'story',
-    'poem',
-    'creative',
-    'brainstorm',
-    'imagine',
-    'fiction',
-    'narrative',
-    'character',
-    'plot',
-    'dialogue',
-    'write a song',
-    'compose',
-    'artistic',
-    'metaphor',
-    'allegory',
-  ],
-
-  questionComplexity: [
-    'how does x relate to y',
-    'what are the implications',
-    'why would',
-    'what happens if',
-    'under what conditions',
-    'how would you approach',
-    'what is the relationship between',
-  ],
-
-  imperativeVerbs: [
-    'build',
-    'create',
-    'update',
-    'deploy',
-    'send',
-    'check',
-    'run',
-    'install',
-    'configure',
-    'set up',
-    'launch',
-    'publish',
-    'submit',
-    'execute',
-    'start',
-    'stop',
-    'restart',
-    'delete',
-    'remove',
-    'review',
-    'optimize',
-    'scan',
-  ],
-
-  outputFormat: [
-    'as json',
-    'in json',
-    'as yaml',
-    'in yaml',
-    'as csv',
-    'markdown',
-    'as a table',
-    'in a table',
-    'as xml',
-    'formatted as',
-    'output as',
-    'return as',
-    'in the format',
-  ],
-
-  domainSpecificity: [
-    'p-value',
-    'confidence interval',
-    'regression',
-    'hipaa',
-    'gdpr',
-    'sec filing',
-    'tort',
-    'liability',
-    'fiduciary',
-    'amortization',
-    'eigenvalue',
-    'fourier transform',
-    'bayesian',
-    'posterior',
-    'genome',
-    'phenotype',
-    'pharmacokinetics',
-    'distribution',
-    'probability',
-    'statistics',
-    'calculate',
-  ],
-
-  agenticTasks: [
-    'triage',
-    'audit',
-    'investigate',
-    'monitor',
-    'orchestrate',
-    'coordinate',
-    'schedule',
-    'prioritize',
-    'delegate',
-    'batch process',
-    'scan all',
-    'check all',
-    'review all',
-    'update all',
-    'migrate',
-    'remediation',
-  ],
-
-  relay: [
-    'forward to',
-    'escalate',
-    'transfer to',
-    'pass along',
-    'relay',
-    'just say',
-    'tell them',
-    'send this to',
-    'notify',
-    'ping',
-    'acknowledge',
-    'confirm receipt',
-    'mark as read',
-  ],
-
-  webBrowsing: [
-    'browse',
-    'navigate',
-    'visit',
-    'open url',
-    'web page',
-    'website',
-    'click on',
-    'scroll',
-    'screenshot',
-    'html',
-    'dom',
-    'css selector',
-    'xpath',
-    'scrape',
-    'web search',
-    'fetch url',
-    'crawl',
-    'url',
-    'http',
-    'https',
-    'link',
-    'page',
-    'site',
-    'download',
-    'bookmark',
-    'go to',
-    'open this',
-    'look up',
-    'fetch',
-    'google',
-    'tab',
-    'domain',
-    'whois',
-    'rss',
-    'sitemap',
-  ],
-
-  dataAnalysis: [
-    'csv',
-    'dataframe',
-    'pandas',
-    'numpy',
-    'chart',
-    'plot',
-    'visualization',
-    'aggregate',
-    'pivot',
-    'correlation',
-    'data analysis',
-    'dataset',
-    'spreadsheet',
-    'histogram',
-    'scatter plot',
-    'heatmap',
-    'dashboard',
-    'metric',
-    'metrics',
-    'kpi',
-    'funnel',
-    'cohort',
-    'segment',
-    'median',
-    'percentile',
-    'variance',
-    'z-score',
-    'chi-square',
-    'anova',
-    'f-statistic',
-    'covariance',
-    'interquartile',
-    'geometric mean',
-    'trend',
-    'forecast',
-    'time series',
-    'seasonal',
-    'demographic',
-    'survey',
-    'sankey',
-    'treemap',
-    'waterfall chart',
-    'pareto',
-    'roi',
-    'churn',
-    'retention',
-    'attrition',
-    'burn rate',
-    'conversion rate',
-    'engagement',
-    'customer lifetime value',
-    'growth rate',
-    'analyze',
-    'compute',
-    'visualize',
-  ],
-
-  imageGeneration: [
-    'generate image',
-    'create image',
-    'draw',
-    'illustration',
-    'render',
-    'photo',
-    'picture',
-    'artwork',
-    'sketch',
-    'logo',
-    'banner',
-    'thumbnail',
-    'wallpaper',
-    'icon design',
-    'visual',
-    'image edit',
-    'inpaint',
-    'outpaint',
-    'upscale',
-    'midjourney',
-    'flux',
-    'ideogram',
-    'leonardo.ai',
-    'leonardo ai',
-    'adobe firefly',
-    'firefly',
-    'gpt-4o image',
-    'stable diffusion',
-    'comfyui',
-    'img2img',
-    'txt2img',
-    'design',
-    'poster',
-    'flyer',
-    'mockup',
-    'avatar',
-    'mascot',
-    'sticker',
-    'collage',
-    'meme',
-    'pattern',
-    'portrait',
-    'invitation',
-    'pixel art',
-    'concept art',
-    'badge',
-    'label',
-    'cover',
-    'graphic',
-    'infographic',
-    'diagram',
-    'template',
-    'certificate',
-    'wireframe',
-    'image',
-    'icon',
-    'painting',
-    'photograph',
-    'gif',
-    'mood board',
-    'map',
-    'character',
-    'texture',
-    'background',
-    'shot',
-    'neon',
-    'botanical',
-  ],
-
-  videoGeneration: [
-    'generate video',
-    'create video',
-    'animation',
-    'animate',
-    'video edit',
-    'clip',
-    'footage',
-    'motion graphics',
-    'timelapse',
-    'transition',
-    'subtitle',
-    'storyboard',
-    'keyframe',
-    'render video',
-    'sora',
-    'runway',
-    'kling',
-    'pika',
-    'veo',
-    'synthesia',
-    'heygen',
-    'capcut',
-    'opusclip',
-    'higgsfield',
-    'text to video',
-    'image to video',
-    'video',
-    'reel',
-    'shorts',
-    'trailer',
-    'montage',
-    'voiceover',
-    'lip sync',
-    'slow motion',
-    'color grading',
-    'green screen',
-    'cinematic',
-    'looping',
-    'captions',
-    'speed ramp',
-    'stop motion',
-    'explainer',
-    'promo',
-    'testimonial',
-    'tutorial video',
-    'demo video',
-    'film',
-    'recording',
-    'webinar',
-    'camera',
-    'scene',
-    'frame',
-    'make a video',
-    'create a video',
-    'produce a video',
-    'generate a video',
-    'video for',
-    'youtube',
-    'tiktok video',
-    'instagram reel',
-    'video content',
-    'video template',
-    'video presentation',
-    'video portfolio',
-    'video greeting',
-    'video resume',
-    'screen recording',
-    'video loop',
-    'video ad',
-    'video post',
-    'video story',
-  ],
-
-  socialMedia: [
-    'tweet',
-    'post',
-    'caption',
-    'hashtag',
-    'instagram',
-    'tiktok',
-    'linkedin',
-    'facebook',
-    'twitter',
-    'social media',
-    'followers',
-    'engagement',
-    'viral',
-    'repost',
-    'thread',
-    'schedule post',
-    'content calendar',
-    'influencer',
-    'audience',
-    'reach',
-    'hootsuite',
-    'buffer',
-    'sprout social',
-    'later',
-    'feedhive',
-    'brand24',
-    'bluesky',
-    'threads',
-    'youtube shorts',
-    'reel',
-    'story',
-    'carousel',
-    'bio',
-    'giveaway',
-    'brand',
-    'community',
-    'content',
-    'poll',
-    'campaign',
-    'spotlight',
-    'milestone',
-    'evergreen',
-    'announcement',
-    'DM',
-    'social ad',
-    'trending',
-    'hook',
-    'pin',
-    'pinterest',
-    'write a post',
-    'draft a post',
-    'create a post',
-    'social strategy',
-    'social content',
-    'social ad',
-    'new feature',
-    'testimonial',
-    'user generated',
-    'cross platform',
-    'social listening',
-    'social coverage',
-    'social report',
-    'write copy',
-    'engagement rate',
-    'share of voice',
-  ],
-
-  emailManagement: [
-    'email',
-    'inbox',
-    'reply to',
-    'forward email',
-    'draft email',
-    'compose email',
-    'subject line',
-    'newsletter',
-    'outreach',
-    'follow up',
-    'mail merge',
-    'unsubscribe',
-    'cold email',
-    'email template',
-    'cc',
-    'bcc',
-    'attachment',
-    'signature',
-    'gmail',
-    'outlook',
-    'superhuman',
-    'spark mail',
-    'shortwave',
-    'mailchimp',
-    'sendgrid',
-    'resend',
-    'drip campaign',
-  ],
-
-  calendarManagement: [
-    'schedule',
-    'meeting',
-    'calendar',
-    'appointment',
-    'availability',
-    'time slot',
-    'reschedule',
-    'invite',
-    'recurring',
-    'agenda',
-    'reminder',
-    'block time',
-    'free slot',
-    'booking',
-    'standup',
-    'sync calendar',
-    'calendly',
-    'cal.com',
-    'reclaim',
-    'reclaim.ai',
-    'clockwise',
-    'motion',
-    'google calendar',
-    'focus time',
-    'cancel',
-    'room',
-    'buffer',
-    'timezone',
-    'offsite',
-    'sprint planning',
-    'retrospective',
-    'itinerary',
-    'deadline',
-    'morning routine',
-    'lunch break',
-    'office hours',
-    'vacation',
-    'holiday',
-    'workout',
-    'dentist',
-    'haircut',
-    'date night',
-    'meal prep',
-    'commute',
-    'flight',
-    'vet',
-    'conference',
-    'training',
-    'onboarding',
-    'catch up',
-    'book a',
-    'move the',
-    'find a time',
-    'check my',
-    'add to calendar',
-    'set up a',
-    'plan a',
-    'weekly sync',
-    'daily standup',
-    'team building',
-    'time block',
-    'next available',
-    'side project',
-    'graduation',
-    'birthday party',
-    'family visit',
-    'networking event',
-    'tax filing',
-    'study group',
-  ],
-
-  trading: [
-    'buy',
-    'sell',
-    'trade',
-    'stock',
-    'crypto',
-    'bitcoin',
-    'ethereum',
-    'forex',
-    'portfolio',
-    'market',
-    'ticker',
-    'candle',
-    'bullish',
-    'bearish',
-    'stop loss',
-    'take profit',
-    'leverage',
-    'margin',
-    'order book',
-    'moving average',
-    'rsi',
-    'macd',
-    'support and resistance',
-    'breakout',
-    'dividend',
-    'hedge',
-    'kalshi',
-    'polymarket',
-    'robinhood',
-    'interactive brokers',
-    'webull',
-    'tradingview',
-    'coinbase',
-    'kraken',
-    'binance',
-    'bitget',
-    'prediction market',
-    'event contract',
-    'solana',
-    'defi',
-    'shares',
-    'options',
-    'futures',
-    'etf',
-    'price',
-    'earnings',
-    'yield',
-    'pe ratio',
-    'volatility',
-    'vix',
-    'fibonacci',
-    'position',
-    'entry',
-    'short interest',
-    'staking',
-    'arbitrage',
-    'sharpe ratio',
-    'alpha',
-    'beta',
-    'drawdown',
-    'swing trade',
-    'day trading',
-    'limit order',
-    'market order',
-    'fomc',
-    'cpi',
-    'sec',
-    'insider trading',
-    'commodity',
-    'bond',
-    'fund',
-    'exchange',
-    'spread',
-    'order',
-    'position size',
-    'risk reward',
-    'dark pool',
-    'kelly criterion',
-    'momentum',
-    'sector rotation',
-    'growth stock',
-    'economic calendar',
-    'stock price',
-    'current price',
-    'best entry',
-    'place a',
-    'place an order',
-    'check my portfolio',
-    'top gaining',
-    'pre market',
-    'after hours',
-    'earnings call',
-    'earnings date',
-    'trading journal',
-    'risk reward ratio',
-    'position sizing',
-    'copy trading',
-    'paper trading',
-    'candlestick',
-    'oversold',
-    'overbought',
-    'resistance level',
-    'support level',
-    'price action',
-    'moving average crossover',
-    'death cross',
-    'golden cross',
-    'fear and greed',
-    'funding rate',
-    'put call',
-    'implied volatility',
-    'options chain',
-    '52 week',
-    'all time high',
-    'market cap',
-    'dollar cost averaging',
-    'compound interest',
-    'total return',
-    'annual return',
-    'expense ratio',
-    'analyze the chart',
-    'analyze the pattern',
-    'analyze the trend',
-    'analyze the market',
-    'technical analysis',
-    'fundamental analysis',
-    'price target',
-    'market sentiment',
-    'whale activity',
-    'accumulation',
-    'distribution',
-    'cup and handle',
-    'head and shoulders',
-    'double bottom',
-    'double top',
-    'wedge pattern',
-    'flag pattern',
-    'pennant',
-    'volume profile',
-    'retracement',
-  ],
-};
+export { DEFAULT_KEYWORDS } from './keywords';
 
 export const DEFAULT_CONFIG: ScorerConfig = {
   dimensions: [
@@ -1012,18 +64,8 @@ export const DEFAULT_CONFIG: ScorerConfig = {
       keywords: DEFAULT_KEYWORDS.agenticTasks,
     },
     { name: 'relay', weight: 0.02, direction: 'down', keywords: DEFAULT_KEYWORDS.relay },
-    {
-      name: 'webBrowsing',
-      weight: 0,
-      direction: 'up',
-      keywords: DEFAULT_KEYWORDS.webBrowsing,
-    },
-    {
-      name: 'dataAnalysis',
-      weight: 0,
-      direction: 'up',
-      keywords: DEFAULT_KEYWORDS.dataAnalysis,
-    },
+    { name: 'webBrowsing', weight: 0, direction: 'up', keywords: DEFAULT_KEYWORDS.webBrowsing },
+    { name: 'dataAnalysis', weight: 0, direction: 'up', keywords: DEFAULT_KEYWORDS.dataAnalysis },
     {
       name: 'imageGeneration',
       weight: 0,
@@ -1036,12 +78,7 @@ export const DEFAULT_CONFIG: ScorerConfig = {
       direction: 'up',
       keywords: DEFAULT_KEYWORDS.videoGeneration,
     },
-    {
-      name: 'socialMedia',
-      weight: 0,
-      direction: 'up',
-      keywords: DEFAULT_KEYWORDS.socialMedia,
-    },
+    { name: 'socialMedia', weight: 0, direction: 'up', keywords: DEFAULT_KEYWORDS.socialMedia },
     {
       name: 'emailManagement',
       weight: 0,
@@ -1054,12 +91,7 @@ export const DEFAULT_CONFIG: ScorerConfig = {
       direction: 'up',
       keywords: DEFAULT_KEYWORDS.calendarManagement,
     },
-    {
-      name: 'trading',
-      weight: 0,
-      direction: 'up',
-      keywords: DEFAULT_KEYWORDS.trading,
-    },
+    { name: 'trading', weight: 0, direction: 'up', keywords: DEFAULT_KEYWORDS.trading },
     { name: 'tokenCount', weight: 0.05, direction: 'up' },
     { name: 'nestedListDepth', weight: 0.03, direction: 'up' },
     { name: 'conditionalLogic', weight: 0.03, direction: 'up' },
